refactor: move function to hooks folder
